requirements.txt
flask==3.1.0 
flask-cors==5.0.1
pyswisseph==2.10.3.2
geopy==2.4.1
timezonefinder==6.5.9
pytz==2025.2
numpy==1.26.4
pandas==2.2.3
gunicorn==23.0.0